backend/requirements.txt
gunicorn
uvicorn
fastapi
httpx
python-dotenv
pydantic
